Use "descriptor" and "RTT" more precisely (#4)
Clarify that an "RTT" is specifically the engine-managed type information inside
a custom descriptor. Update the overview to use the terminology more uniformly
and update the new instruction names to use `_desc` instead of `_rtt` as well.

Also take the opportunity to update the text format for `describes` and
`descriptor` clauses to match the proposed binary format.