Describe a "distinguished value" API for JS prototypes
The JS interop design previously described in the overview depended on
the identities of prototypes attached to custom descriptors being
mutable. Implementer feedback from both SpiderMonkey and V8 was that the
identities need to be immutable, so that design was not feasible.
Although the conversation about what the final design should look like
is not over, replace the old design in the overview with one that meets
more of our requirements.