Generalize the binary shorthand for exact references
Extend the pattern of using `0x62 ht:absheaptype` as a shorthand for
nullable reference types beyond just abstract heaptypes. Instead, use
`0x62 ht:heaptype` to mean a nullable exact reference for any heap type,
replacing the previous encoding.

Keep the `(exact anyref)` text shorthand for abstract heap types, but do
not extend it to other heap types since it is generally not as readable
as an explicitly spelled out type.